support floating point x values. also: --nline x0 x1 N

+++ python_rewrite_claude/cxx/trace.h
@@ -9,7 +9,7 @@ struct TraceOptions {
     int direction;
     int nlines;
     int nturns;
-    std::vector<int> lines;
+    std::vector<float> lines;
 };
 
 void trace_field_lines(const FieldData& field, const TraceOptions& options);
